Refactor view -> select Refactor list -> view Refactor select -> locate
