+++ src/main/java/seedu/address/logic/commands/SelectCommand.java
@@ -2,28 +2,28 @@ package seedu.address.logic.commands;
 
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.core.Messages;
+import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.commons.events.ui.JumpToListRequestEvent;
 import seedu.address.model.task.ReadOnlyTask;
-import seedu.address.commons.core.UnmodifiableObservableList;
 
+//@@author A0135767U
 /**
- * Selects a task identified using it's last displayed index from the SmartyDo.
+ * Lists all tasks in the SmartyDo to the user.
  */
 public class SelectCommand extends Command {
 
-    public final int targetIndex;
-
     public static final String COMMAND_WORD = "select";
-
-    public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Selects the task identified by the index number used in the last task listing.\n"
-            + "Parameters: INDEX (must be a positive integer)\n"
+    
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Shows details of a specified task.\n"
+            + "Parameters: INDEX \n"
             + "Example: " + COMMAND_WORD + " 1";
 
-    public static final String MESSAGE_SELECT_TASK_SUCCESS = "Selected Task: %1$s";
+    public static final String MESSAGE_SUCCESS = "Selected Task #%1$s";
 
+    public final int targetIndex;
+    
     public SelectCommand(int targetIndex) {
-        this.targetIndex = targetIndex;
+    	this.targetIndex = targetIndex;
     }
 
     @Override
@@ -37,7 +37,7 @@ public class SelectCommand extends Command {
         }
 
         EventsCenter.getInstance().post(new JumpToListRequestEvent(targetIndex - 1));
-        return new CommandResult(String.format(MESSAGE_SELECT_TASK_SUCCESS, targetIndex));
-
+        return new CommandResult(String.format(MESSAGE_SUCCESS, targetIndex));
+        
     }
 }
